resolved error regarding too many redirects from login page

+++ ShoppingCartWebApp/Controllers/LogoutController.cs
@@ -22,13 +22,15 @@ namespace ShoppingCartWebApp.Controllers
         {
             Response.Cookies.Delete("SessionId");
             Response.Cookies.Delete("Username");
+            if (Request.Cookies["SessionId"] != null)
+            {
+                Guid sessionId = Guid.Parse(Request.Cookies["sessionId"]);
+                Session session = dbContext.Sessions.FirstOrDefault(x =>
+                    x.Id == sessionId.ToString()
+                );
 
-            Guid sessionId = Guid.Parse(Request.Cookies["sessionId"]);
-            Session session = dbContext.Sessions.FirstOrDefault(x =>
-                x.Id == sessionId.ToString()
-            );
-
-            db.DeleteSessionData(session);
+                db.DeleteSessionData(session);
+            }
             
 
             return RedirectToAction("Index", "Login");
